refactor: 계산함수 분리 및 key관련 warning제거
